Doh! What did I do here? File was locked by existing server instance and I was trying to delete it. Also, this happened every start, which was not the intent here. Now test for the existance of the named executable and simply copy if the source file exists and we renamed the file in the config.

+++ BedrockService/BedrockServiceWrapper.cs
@@ -236,14 +236,11 @@ namespace BedrockService
                 }
                 foreach (var server in bedrockServers)
                 {
-                    if (server.ServerConfig.BedrockServerExeName != "bedrock_server.exe" && File.Exists(server.ServerConfig.BedrockServerExeLocation + "bedrock_server.exe"))
+                    if (server.ServerConfig.BedrockServerExeName != "bedrock_server.exe" && File.Exists(server.ServerConfig.BedrockServerExeLocation + "bedrock_server.exe") && !File.Exists(server.ServerConfig.BedrockServerExeLocation + server.ServerConfig.BedrockServerExeName))
                     {
-                        if (File.Exists(server.ServerConfig.BedrockServerExeLocation + server.ServerConfig.BedrockServerExeName))
-                        {
-                            File.Delete(server.ServerConfig.BedrockServerExeLocation + server.ServerConfig.BedrockServerExeName);
-                        }
                         File.Copy(server.ServerConfig.BedrockServerExeLocation + "bedrock_server.exe", server.ServerConfig.BedrockServerExeLocation + server.ServerConfig.BedrockServerExeName);
                     }
+
                     if (!File.Exists(server.ServerConfig.BedrockServerExeLocation + server.ServerConfig.BedrockServerExeName))
                     {
                         if (File.Exists($@"{server.ServerConfig.BedrockServerExeLocation}FileList.ini"))
@@ -266,7 +263,8 @@ namespace BedrockService
                         }
                         catch (Exception e)
                         {
-                            Console.WriteLine($"ERROR: Got zipfile exception! {e.Message}");
+                            Console.WriteLine($"ERROR: Got zipfile exception! {e.Message}.\n!!Deleting entire directory!! Press Enter to continue...");
+                            Console.ReadLine();
                             DeleteFilesRecursively(new DirectoryInfo(server.ServerConfig.BedrockServerExeLocation));
                             ValidSettingsCheck();
                         }
